refactor: 역할에 따라 패키지 이동
Model View Controller의 역할에 따라 패키지를 분리
