+++ src/main/java/racingcar/ApplicationMain.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.controller.RacingGame;
+
 public class ApplicationMain {
 
     public static void main(String[] args) {
