+++ src/main/java/racingcar/controller/CarStatusDto.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.controller;
 
 public class CarStatusDto {
