+++ src/main/java/racingcar/model/CarCondition.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import java.util.Random;
