+++ src/main/java/racingcar/model/CarNameSeparator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import java.util.List;
