+++ src/main/java/racingcar/model/Condition.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 public interface Condition {
