+++ src/main/java/racingcar/model/nickname/CanNotBlankNicknameException.java
@@ -1,4 +1,4 @@
-package racingcar.nickname;
+package racingcar.model.nickname;
 
 public class CanNotBlankNicknameException extends RuntimeException {
