+++ src/main/java/racingcar/model/nickname/CanNotNullNicknameException.java
@@ -1,4 +1,4 @@
-package racingcar.nickname;
+package racingcar.model.nickname;
 
 public class CanNotNullNicknameException extends RuntimeException {
