+++ src/test/java/racingcar/model/CarNameSeparatorTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import org.junit.jupiter.api.Test;
