+++ src/test/java/racingcar/model/GoodCondition.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 public class GoodCondition implements Condition {
